Menu bar action to file a GitHub issue (bazelbuild/intellij PR import #1061)
This adds a new menu bar reference to file a GitHub issue directly from the IDE.

When the user clicks on the reference, the action obtains relevant version information and opens the new issue page and pre-fills the information. The user then fills in the questionnaire and manually submits the issue.

Sample issue content:

```
#### Description of the issue. Please be specific.

#### What's the simplest set of steps to reproduce this issue? Please provide an example project, if possible.

#### Version information
IdeaCommunity: 2019.1.2
Platform: Linux 4.19.37-5rodete1-amd64
Bazel plugin: 9999
Bazel: 0.28.1
```

Closes https://github.com/bazelbuild/intellij/pull/1061

PiperOrigin-RevId: 264646279