Use ManifestParser.ParsedManifest in BlazeAndroidDeployInfo
Currently BlazeAndroidDeployInfo is a wrapper around AndroidDeployInfo
proto and contains it's own extraction logic for manifest values. This
is unecessarily complex and inflexible, so this CL extracts the logic
out of BlazeAndroidDeployInfo so it's now just a data-class. This added
flexibility also aids the upcoming android_instrumentation_test support
feature, where manifests values are extracted differently depending on
the build step.

Related application ID providers are updated and new unit tests for the
application ID providers are added.

PiperOrigin-RevId: 275288802