Reindex for external libraries after 5 seconds of no new files being added.
Previously we relied on the delay of the indexing itself (usually 5 seconds apart) to batch the indexing of new files, but this could result in continuous indexing until files stop being added. We will change the order: wait for files to stop being added first, then reindex all at once. At the worst case, this will be identical to the previous behavior, if the user just happens to have new files added exactly 5 seconds apart, but this is very unlikely.

PiperOrigin-RevId: 288712326