Automated rollback of commit ca40ade81d1daf30b508a81edc54130a05e5dc25.
*** Reason for rollback ***

Can use ApplicationInfo.getBuild instead

*** Original change description ***

Use sdkcompat to determine if DisableDockerPlugin should be deactivated

The code relied on having access to PluginManagerCore.BUILD_NUMBER, which has
been removed. Rather than exposing the BUILD_NUMBER via reflection, this CL
just exposes a #shouldDisableDockerPlugin that is customized per API.

We may still have to somehow expose the BUILD_NUMBER functionality as some
integration tests may rely on it, but that'll be fixed when we come to that
failure.

PiperOrigin-RevId: 314621478