Hack to prevent IntelliJ from sending defaultConfig to typescript service.
If we have a single tsconfig file, IntelliJ will send a defaultConfig to
the language service, which will override the isUseSingleInferredProject
that we set and cause crashes in the language service when it tries to
watch certain directories. We'll return an empty list here to fool
IntelliJ into not sending the defaultConfig.

This is extremely hacky. The proper fix should likely be somewhere in
the tsconfig.json or in the typescript service.

PiperOrigin-RevId: 323041723