Add check for AndroidFacet in module verification
The AndroidFacet is required for modules used to execute
run configurations, in this case the ".workspace" module.

Add a user visible error that asks for a resync if the
AndroidFacet can't be found as it indicates incomplete or
bad previous sync. Also add a quick fix button to trigger
an incremental sync next to the error message.

PiperOrigin-RevId: 333213437