Cache library dependencies for the workspace module
In AS 4.1, the `AndroidResolveScopeEnlarger` ends up getting called quite often.
It in turn invokes `BlazeModuleSystem.getResolvedDependentLibraries`. This method
is quite slow as it relies on calling `BlazeLibraryCollector.getLibraries` which
tries to filter all libraries by passing them through a glob.

To work around this, we cache the result of the computation using SyncCache (which
takes care of clearing the cache after a sync is complete).

PiperOrigin-RevId: 340955838