Move away from internal PlatformProjectOpenProcessor.openExistingProject()
ProjectManagerEx.openProject() is a replacement for this API in 2020.2 and
is considered to be a public API (confirmed with JetBrains). The previous
method call had three arguments. Internally, it also just delegated
to ProjectManagerEx.openProject() when the first two parameters were the
same (which is true for our case).

Fortunately, we already had the call in sdkcompat and some other part of it
changed in 2020.2. Hence, we can just piggy-back on the compat code and
move to the public API from 2020.2+ on.

PiperOrigin-RevId: 357990677