Fix AbstractMethodError for CLion users
Users see the exception when they try to open a project which contains a
CMakeLists.txt file.

Apparently from 2020.1 on, implementations of EditorNotifications throw an
exception if they don't override the createNotificationPanel() method which
additionally takes a Project parameter. The old override of the method without
the Project parameter is not necessary anymore.

PiperOrigin-RevId: 359794241